% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoresid-package.R
\docType{package}
\name{autoresid-package}
\alias{autoresid}
\alias{autoresid-package}
\title{autoresid: Auto (Distance) Covariance Tests for Residuals}
\description{
The autoresid package performs model diagnostics by checking for dependence among estimated residuals.
}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
