% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{autoresid._Arima_fit_impl}
\alias{autoresid._Arima_fit_impl}
\alias{autoresid.Arima_fit_impl}
\alias{autoresid.Arima}
\title{Autoresid methods for ARIMA models}
\usage{
\method{autoresid}{`_Arima_fit_impl`}(object, new_data, outcome, ...)

\method{autoresid}{Arima_fit_impl}(object, new_data, outcome, ...)

\method{autoresid}{Arima}(object, new_data, outcome, ...)
}
\arguments{
\item{object}{a fitted arima or garch model}

\item{new_data}{a tibble of univariate time series on which residuals are to
be calculated.}

\item{outcome}{name of column to fit data to}

\item{...}{Currently unused.}
}
\description{
Autoresid methods for ARIMA models
}
