% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{autoresid.lm}
\alias{autoresid.lm}
\alias{autoresid._lm}
\title{Autoresid methods for lm models}
\usage{
\method{autoresid}{lm}(object, new_data, outcome, ...)

\method{autoresid}{`_lm`}(object, new_data, outcome, ...)
}
\arguments{
\item{object}{a fitted arima or garch model}

\item{new_data}{a tibble of univariate time series on which residuals are to
be calculated.}

\item{outcome}{name of column to fit data to}

\item{...}{Currently unused.}
}
\description{
It is assumed that \code{new_data} has all the predictor variables. However the
\code{outcome} needs to be specified in this implementation.
}
\examples{
lin_mod <- lm(mpg ~ cyl, data = mtcars)
autoresid(lin_mod, head(mtcars), "mpg")
}
