% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metrics_smp_spl.R
\name{collect_metrics.smp_spl_results}
\alias{collect_metrics.smp_spl_results}
\title{Collect metrics for sample splitting}
\usage{
\method{collect_metrics}{smp_spl_results}(x)
}
\arguments{
\item{x}{The results of \code{\link[tune:tune_grid]{tune_grid()}}, \code{\link[tune:tune_bayes]{tune_bayes()}}, \code{\link[tune:fit_resamples]{fit_resamples()}},
or \code{\link[tune:last_fit]{last_fit()}}. For \code{\link[tune:collect_predictions]{collect_predictions()}}, the control option \code{save_pred = TRUE} should have been used.}
}
\description{
Collect metrics for sample splitting
}
