% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.R
\name{extract_model}
\alias{extract_model}
\alias{extract_model.Arima}
\alias{extract_model.Arima_fit_impl}
\alias{extract_model._Arima_fit_impl}
\alias{extract_model.fGarch}
\alias{extract_model._fGarch}
\title{Extracts available models}
\usage{
extract_model(object, ...)

\method{extract_model}{Arima}(object, ...)

\method{extract_model}{Arima_fit_impl}(object, ...)

\method{extract_model}{`_Arima_fit_impl`}(object, ...)

\method{extract_model}{fGarch}(object, ...)

\method{extract_model}{`_fGarch`}(object, ...)
}
\arguments{
\item{object}{an object to extract model from.}

\item{...}{unused.}
}
\description{
Internal generic function to extract and return objects that have been
registered with this package.
}
