% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_splits.R
\name{fit_splits}
\alias{fit_splits}
\title{Fit multiple sample splits}
\usage{
fit_splits(
  object,
  data,
  num_analysis,
  num_assessment,
  use_fit_resamples = FALSE,
  ...,
  metrics = NULL,
  control = tune::control_resamples()
)
}
\arguments{
\item{object}{A \code{parsnip} model specification or a \code{\link[workflows:workflow]{workflows::workflow()}}.
No tuning parameters are allowed.}

\item{data}{A data frame of predictors and outcomes to use when fitting the
workflow.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_assessment}{A positive integer specifying number of observations
for goodness-of-fit tests.}

\item{use_fit_resamples}{If \code{TRUE}, uses \code{\link[tune:fit_resamples]{tune::fit_resamples()}}.}

\item{...}{Currently unused.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}, or \code{NULL} to compute a standard
set of metrics.}

\item{control}{A \code{\link[tune:control_resamples]{control_resamples()}} object used to fine tune the resampling
process.}
}
\description{
Fits model according to sample splitting on a grid of values determined by
\code{num_analysis} and \code{num_assessment}, where each tuple determines the number
of analysis and assessment sets.
}
