% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-residuals.R
\name{residuals._Arima_fit_impl}
\alias{residuals._Arima_fit_impl}
\title{Extract fitted residuals from arima models}
\usage{
\method{residuals}{`_Arima_fit_impl`}(object, type = c("innovation", "response", "regression"), h = 1, ...)
}
\arguments{
\item{object}{An object containing a time series model of class \code{ar},
\code{Arima}, \code{bats}, \code{ets}, \code{arfima}, \code{nnetar} or
\code{stlm}.
If \code{object} is of class \code{forecast}, then the function will return
\code{object$residuals} if it exists, otherwise it returns the differences between
the observations and their fitted values.}

\item{type}{Type of residual.}

\item{h}{If \code{type='response'}, then the fitted values are computed for
\code{h}-step forecasts.}

\item{...}{Other arguments not used.}
}
\description{
Extract fitted residuals from arima models
}
