% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoresid-package.R
\docType{package}
\name{autoresid-package}
\alias{autoresid-package}
\alias{_PACKAGE}
\title{autoresid: Auto Residuals from Fitted Models}
\description{
The autoresid package returns the residuals from various models including ARIMA and GARCH models.
}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
