% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/autoresid-generic.R, R/garch.R
\name{autoresid._Arima_fit_impl}
\alias{autoresid._Arima_fit_impl}
\alias{autoresid.Arima}
\alias{autoresid}
\alias{autoresid.fGARCH}
\title{Extract fitted residuals from models}
\usage{
\method{autoresid}{`_Arima_fit_impl`}(object, new_data, outcome, ...)

\method{autoresid}{Arima}(object, new_data, outcome, ...)

autoresid(object, new_data, outcome, ...)

\method{autoresid}{fGARCH}(object, new_data, outcome, standardize_resid = FALSE, ...)
}
\arguments{
\item{object}{a fitted arima or garch model}

\item{new_data}{a tibble of univariate time series on which residuals are to
be calculated.}

\item{outcome}{name of column to fit data to}

\item{...}{Currently unused.}

\item{standardize_resid}{single logical. Should the extracted residuals be
stanrdized?}
}
\value{
a tibble of predicted values and residuals.
}
\description{
Extract fitted residuals from models
}
