% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/autoresid-generic.R
\name{autoresid._Arima_fit_impl}
\alias{autoresid._Arima_fit_impl}
\alias{autoresid.Arima}
\alias{autoresid}
\title{Extract fitted residuals from models}
\usage{
\method{autoresid}{`_Arima_fit_impl`}(object, new_data, ...)

\method{autoresid}{Arima}(object, new_data, ...)

autoresid(object, new_data, ...)
}
\arguments{
\item{object}{a fitted arima or garch model}

\item{new_data}{a tibble of univariate time series on which residuals are to
be calculated.}

\item{...}{Currently unused.}
}
\value{
a tibble of predicted values and residuals.
}
\description{
Extract fitted residuals from models
}
