% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{autoresid_arima_impl}
\alias{autoresid_arima_impl}
\title{Autoresid implementation for arima models}
\usage{
autoresid_arima_impl(object, new_data, outcome, ...)
}
\arguments{
\item{object}{a fitted arima or garch model}

\item{new_data}{a \link[tsibble:tsibble-package]{tsibble} of univariate
time series on which residuals are to be calculated.}

\item{outcome}{name of column to fit data to.}

\item{...}{Options passed to methods.}
}
\description{
Autoresid implementation for arima models
}
