% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{autoresid_fGARCH_impl}
\alias{autoresid_fGARCH_impl}
\title{Autoresid methods for GARCH models}
\usage{
autoresid_fGARCH_impl(
  object,
  new_data,
  outcome,
  standardize_resid = FALSE,
  ...
)
}
\arguments{
\item{object}{a fitted arima or garch model}

\item{new_data}{a \link[tsibble:tsibble-package]{tsibble} of univariate
time series on which residuals are to be calculated.}

\item{outcome}{name of column to fit data to.}

\item{standardize_resid}{single logical. Should the extracted residuals be
standardized?}

\item{...}{Options passed to methods.}
}
\description{
Autoresid methods for GARCH models
}
