% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.R, R/extract_outcome.R
\name{extract_model}
\alias{extract_model}
\alias{extract_model.model_fit}
\alias{extract_model.Arima}
\alias{extract_model.Arima_fit_impl}
\alias{extract_model.mdl_ts}
\alias{extract_model.ARIMA}
\alias{extract_model.AR}
\alias{extract_model.fGARCH}
\alias{extract_outcome.mdl_defn}
\title{Extracts available models}
\usage{
extract_model(object, ...)

\method{extract_model}{model_fit}(object, ...)

\method{extract_model}{Arima}(object, ...)

\method{extract_model}{Arima_fit_impl}(object, ...)

\method{extract_model}{mdl_ts}(object, ...)

\method{extract_model}{ARIMA}(object, ...)

\method{extract_model}{AR}(object, ...)

\method{extract_model}{fGARCH}(object, ...)

\method{extract_outcome}{mdl_defn}(object, ...)
}
\arguments{
\item{object}{an object to extract model from.}

\item{...}{unused.}
}
\description{
Internal generic function to extract and return objects that have been
registered with this package. For now only \code{Arima} and \code{fGARCH} objects are
registered.
}
